+++ src/enums/index.ts
@@ -3,6 +3,7 @@
 export * from "./BoundingType.js";
 export * from "./CameraEffectType.js";
 export * from "./CameraStrategyType.js";
+export * from "./CubeLayout.js";
 export * from "./EventType.js";
 export * from "./Keys.js";
 export * from "./LightType.js";